+++ data/src/main/java/com/weit/data/di/PlaceModule.kt
@@ -8,7 +8,7 @@ import com.weit.data.BuildConfig
 import com.weit.data.repository.place.PlaceRepositoryImpl
 import com.weit.data.service.PlaceService
 import com.weit.data.source.ImageDataSource
-import com.weit.data.source.PlaceDateSource
+import com.weit.data.source.PlaceDataSource
 import com.weit.domain.repository.place.PlaceRepository
 import dagger.Module
 import dagger.Provides
@@ -47,12 +47,12 @@ class PlaceModule {
         service: PlaceService,
         sessionToken: AutocompleteSessionToken,
         placesClient: PlacesClient,
-    ): PlaceDateSource = PlaceDateSource(context, service, sessionToken, placesClient)
+    ): PlaceDataSource = PlaceDataSource(context, service, sessionToken, placesClient)
 
     @Singleton
     @Provides
     fun providePlaceRepository(
         @ApplicationContext context: Context,
-        dataSource: PlaceDateSource,
+        dataSource: PlaceDataSource,
     ): PlaceRepository = PlaceRepositoryImpl(dataSource, ImageDataSource(context.contentResolver))
 }
