+++ data/src/main/java/com/weit/data/repository/place/PlaceReviewRepositoryImpl.kt
@@ -4,7 +4,7 @@ import android.content.res.Resources.NotFoundException
 import com.weit.data.model.place.PlaceReviewDTO
 import com.weit.data.model.place.PlaceReviewModification
 import com.weit.data.model.place.PlaceReviewRegistration
-import com.weit.data.source.PlaceReviewDateSource
+import com.weit.data.source.PlaceReviewDataSource
 import com.weit.data.util.exception
 import com.weit.domain.model.exception.ForbiddenException
 import com.weit.domain.model.exception.InvalidRequestException
@@ -33,7 +33,7 @@ import java.util.concurrent.atomic.AtomicBoolean
 import javax.inject.Inject
 
 class PlaceReviewRepositoryImpl @Inject constructor(
-    private val dataSource: PlaceReviewDateSource,
+    private val dataSource: PlaceReviewDataSource,
 ) : PlaceReviewRepository {
 
     private val hasNextReviewByPlaceID = AtomicBoolean(true)
