+++ data/src/main/java/com/weit/data/source/FavoritePlaceDateSource.kt
@@ -26,13 +26,8 @@ class FavoritePlaceDateSource @Inject constructor(
         service.register(favoritePlaceRegistration)
     }
 
-    suspend fun delete(favoritePlaceId: Long): Result<Unit> {
-        val response = service.delete(favoritePlaceId)
-        return if (response.isSuccessful) {
-            Result.success(Unit)
-        } else {
-            Result.failure(handleDeleteError(response))
-        }
+    suspend fun delete(favoritePlaceId: Long): Response<Unit> {
+        return service.delete(favoritePlaceId)
     }
 
     suspend fun isFavoritePlace(placeId: String): Boolean =
@@ -48,20 +43,4 @@ class FavoritePlaceDateSource @Inject constructor(
             lastFavoritePlaceId = info.lastFavoritePlaceId,
         )
 
-    private fun handleDeleteError(response: Response<*>): Throwable {
-        return handleCode(response.code())
-    }
-
-    private fun handleCode(code: Int): Throwable {
-        return when (code) {
-            HTTP_NOT_FOUND -> NotExistCommunityIdOrCommunityCommentsException()
-            HTTP_UNAUTHORIZED -> InvalidTokenException()
-            HTTP_FORBIDDEN -> InvalidPermissionException()
-            HTTP_CONFLICT -> ExistedCommunityIdException()
-            HTTP_BAD_REQUEST -> InvalidRequestException()
-            else -> {
-                UnKnownException()
-            }
-        }
-    }
 }
