+++ data/src/main/java/com/weit/data/source/PlaceReviewDataSource.kt
@@ -11,7 +11,7 @@ import com.weit.domain.model.place.PlaceReviewByUserIdQuery
 import retrofit2.Response
 import javax.inject.Inject
 
-class PlaceReviewDateSource @Inject constructor(
+class PlaceReviewDataSource @Inject constructor(
     private val service: PlaceReviewService,
 ) {
