+++ presentation/src/main/java/com/weit/presentation/ui/profile/lifeshot/LifeShotSelectPlaceFragment.kt
@@ -61,6 +61,9 @@ class LifeShotSelectPlaceFragment :
             handleMenuItem(menu.itemId)
             true
         }
+        binding.tbSelectPlace.setNavigationOnClickListener {
+            findNavController().popBackStack()
+        }
     }
 
     private fun handleMenuItem(itemId: Int) {
